Simulation Output File
This is the dictionary that the simulation outputs. There is ALOT of information in this array but we are only interested in the Pressure data. This is saved as States.P. The volume is saved as States.V. I will need to define a helper function to convert the volume array to crank angle array for one of the plots as well.