Cantera Simulation
This is the main simulation code that needs to be tweaked so that it works with the GUI. The integral results section would also be something that we could include in the output on our GUI. I am also thinking about putting slight variances vs time in the inputs so that it is more realistic to how an actual engine behaves. This could be done easily by adding a random integer on a small scale to each of the inputs. For instance, the amount of fuel injected could be some constant based on throttle position plus random integer since in real life, the amount of fuel injected each cycle is not constant. This would make our plots much more realistic during simulation.