Example Data for plotting Code
This data is real life data and what we will be plotting from our simulation results. I was not able to effectively save the simulation outputs directly. If you use this to code your part it will be simple to swap over to simulation results once all our code is put together.